♻️ Refactor CLI into bin/cli modules with withDaemon and --deep fix
Extract helpers, session naming, search/place, and a shared withDaemon()
wrapper from the monolithic lattices.ts router. Wire --deep as an alias for
--all-sources search, align docs with lattices start vs bare lattices home
screen, and add tests/cli.test.mjs for hash parity and daemon behavior.
