+++ docs/config.md
@@ -152,6 +152,7 @@ Run `lattices init` in your project directory to generate a starter
 | `lattices hud sync`          | Publish all registered HUD actors                 |
 | `lattices search <query>`      | Search windows by title, app, session, OCR       |
 | `lattices search <q> --deep`   | Deep search: index + live terminal inspection    |
+| `lattices search <q> --all`    | Same as `--deep` (all search sources)            |
 | `lattices search <q> --wid`    | Print matching window IDs only (pipeable)        |
 | `lattices place <query> [pos]` | Deep search + focus + tile (default: bottom-right)|
 | `lattices focus <session>`   | Focus a session's window and switch Spaces        |
